requirements.txt
anthropic
openai
google-generativeai
groq
pyyaml
python-dotenv
